WIP compute false despite strategy.
