+++ atl-checker/src/algorithms/game_strategy/mod.rs
@@ -1,9 +1,9 @@
 use crate::algorithms::certain_zero::common::VertexAssignment;
 use crate::algorithms::certain_zero::search_strategy::{SearchStrategy, SearchStrategyBuilder};
 use crate::algorithms::certain_zero::{distributed_certain_zero, CertainZeroResult};
-use crate::algorithms::game_strategy::enforcing::compute_enforcing_strategy;
 use crate::algorithms::game_strategy::error::Error;
 use crate::algorithms::game_strategy::format::PartialStrategyWithFormatting;
+use crate::algorithms::game_strategy::partial::{compute_partial_strategy, PartialStrategy};
 use crate::edg::atledg::pmoves::PartialMove;
 use crate::edg::atledg::vertex::ATLVertex;
 use crate::edg::atledg::ATLDependencyGraph;
@@ -11,9 +11,9 @@ use crate::game_structure::{GameStructure, Player, State};
 use std::collections::HashMap;
 use std::fmt::{Debug, Display, Formatter};
 
-pub mod enforcing;
 pub mod error;
 pub mod format;
+pub mod partial;
 
 #[derive(Debug)]
 pub struct ModelCheckResult {
@@ -61,29 +61,6 @@ pub enum SpecificationProof {
     NoStrategyNeeded,
 }
 
-#[derive(Eq, PartialEq, Clone, Debug)]
-pub struct PartialStrategy {
-    /// The players for which this strategy applies
-    players: Vec<Player>,
-    /// A partial mapping from States to a partial move, where the given players have made
-    /// a specific choice.
-    move_to_pick: HashMap<State, PartialMove>,
-}
-
-impl PartialStrategy {
-    /// Pairs a partial strategy with its game structure, allowing us to print
-    /// the strategy using the names of players and moves as defined by the game structure.
-    pub fn in_context_of<'a, G: GameStructure>(
-        &'a self,
-        game: &'a G,
-    ) -> PartialStrategyWithFormatting<'a, G> {
-        PartialStrategyWithFormatting {
-            strategy: self,
-            game,
-        }
-    }
-}
-
 pub fn compute_game_strategy<G: GameStructure>(
     graph: &ATLDependencyGraph<G>,
     v0: &ATLVertex,
@@ -107,11 +84,13 @@ pub fn compute_game_strategy<G: GameStructure>(
 
     match res {
         VertexAssignment::TRUE if formula.is_enforce() => Ok(SpecificationProof::Strategy(
-            compute_enforcing_strategy(graph, v0, assignments),
+            compute_partial_strategy(graph, v0, assignments),
         )),
         VertexAssignment::TRUE if formula.is_despite() => unimplemented!(),
         VertexAssignment::FALSE if formula.is_enforce() => unimplemented!(),
-        VertexAssignment::FALSE if formula.is_despite() => unimplemented!(),
+        VertexAssignment::FALSE if formula.is_despite() => Ok(SpecificationProof::Strategy(
+            compute_partial_strategy(graph, v0, assignments),
+        )),
         _ if formula.players().is_none() => Ok(SpecificationProof::NoStrategyNeeded),
         _ => panic!("Assignment of v0 is undecided"),
     }
